Migrate to NAPI for rs to ts binding (#7)
* Migrate to NAPI for rs to ts binding

* Migrate to NAPI for rs to ts binding

* Migrate to NAPI for rs to ts binding

* Migrate to NAPI for rs to ts binding

* Update snapshots

* Add tagging step to publish action

* eslint disable

* eslint disable

* only lint in pr

+++ src/parser/toc.rs
@@ -22,7 +22,7 @@ fn extract_toc_recursive(ast: &[NorgAST], toc: &mut Vec<TocEntry>) {
             let id = into_slug(&text);
 
             toc.push(TocEntry {
-                level: *level as usize,
+                level: *level as u32,
                 title: text,
                 id,
             });
